+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 - Improved styles of fonts and buttons
 - Added Release feature
+- Improved backlog management by adding release cards to allow easy assigning isssues to releases and filter for releases
 - Bumped Go to 1.25.10 (fixing high security issue in net package)
 - Fixed missing project_id constraint in labels database table and added migration code for existing databases
 - Added a clean reset of project-scoped fields label, release and status, when changing the issue project